// server/src/types/api_response_data.interface.ts
type UrlString = string;
type ImageUrlString = UrlString;

interface IAnswerMessage {
  msg: string;
  preview_url?: ImageUrlString;
}

interface IMealItem {
  name: string;
  link: UrlString; // url
  preview_url: ImageUrlString; // url to image
  cooking_time: string;
  difficulty: string;
}

export interface IApiResponse {
  answer_message: IAnswerMessage;
  suggestions?: string[];
  meal_list?: IMealItem[];
}